requirements.txt
# 기존 패키지
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
youtube-transcript-api>=0.6.1
anthropic>=0.18.1
pandas>=2.0.0
openpyxl>=3.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Whisper 음성 인식 (선택사항)
openai-whisper>=20231117
yt-dlp>=2024.0.0
ffmpeg-python>=0.2.0

# 주의: FFmpeg가 시스템에 설치되어 있어야 합니다
# Windows: https://www.gyan.dev/ffmpeg/builds/
# Mac: brew install ffmpeg
# Linux: sudo apt install ffmpeg